Carve out :vector_value separately from :value
This is just relocating some code; no functional changes.
