+++ drake/systems/framework/BUILD
@@ -26,6 +26,7 @@ drake_cc_library(
         ":system",
         ":value",
         ":vector",
+        ":vector_value",
     ],
 )
 
@@ -62,12 +63,23 @@ drake_cc_library(
 )
 
 drake_cc_library(
+    name = "vector_value",
+    hdrs = ["vector_value.h"],
+    deps = [
+        ":value",
+        ":vector",
+        "//drake/common",
+    ],
+)
+
+drake_cc_library(
     name = "value_checker",
     srcs = [],
     hdrs = ["value_checker.h"],
     visibility = [":__subpackages__"],
     deps = [
         ":value",
+        ":vector_value",
         "//drake/common",
         "//drake/common:nice_type_name",
     ],
@@ -170,6 +182,7 @@ drake_cc_library(
         ":value",
         ":value_checker",
         ":vector",
+        ":vector_value",
         "//drake/common",
     ],
 )
@@ -243,6 +256,7 @@ drake_cc_library(
     visibility = [":__subpackages__"],
     deps = [
         ":value",
+        ":vector_value",
         ":vector",
         "//drake/common",
     ],
@@ -514,6 +528,13 @@ drake_cc_googletest(
     deps = [
         ":value",
         "//drake/common",
+    ],
+)
+
+drake_cc_googletest(
+    name = "vector_value_test",
+    deps = [
+        ":vector_value",
         "//drake/systems/framework/test_utilities",
     ],
 )
