+++ drake/systems/framework/test/vector_value_test.cc
@@ -1,0 +1,94 @@
+#include "drake/systems/framework/vector_value.h"
+
+#include <gtest/gtest.h>
+
+#include "drake/systems/framework/basic_vector.h"
+#include "drake/systems/framework/test_utilities/my_vector.h"
+
+namespace drake {
+namespace systems {
+namespace {
+
+using MyVector2d = MyVector<2, double>;
+
+GTEST_TEST(VectorValueTest, Access) {
+  VectorValue<double> value(BasicVector<double>::Make({1, 2, 3}));
+  EXPECT_EQ(1, value.get_value()->get_value().x());
+  EXPECT_EQ(2, value.get_value()->get_value().y());
+  EXPECT_EQ(3, value.get_value()->get_value().z());
+}
+
+GTEST_TEST(VectorValueTest, CopyConstructor) {
+  VectorValue<double> value(BasicVector<double>::Make({1, 2, 3}));
+
+  VectorValue<double> other_value(value);
+  EXPECT_EQ(1, other_value.get_value()->get_value().x());
+  EXPECT_EQ(2, other_value.get_value()->get_value().y());
+  EXPECT_EQ(3, other_value.get_value()->get_value().z());
+}
+
+GTEST_TEST(VectorValueTest, CopyConstructorSubclass) {
+  VectorValue<double> value(MyVector2d::Make(1, 2));
+  VectorValue<double> other_value(value);
+  value.get_value()->set_value(Eigen::Vector2d(3, 4));
+
+  const MyVector2d* const vector =
+      dynamic_cast<const MyVector2d*>(other_value.get_value());
+  ASSERT_NE(vector, nullptr);
+  EXPECT_EQ(vector->get_value()(0), 1);
+  EXPECT_EQ(vector->get_value()(1), 2);
+}
+
+GTEST_TEST(VectorValueTest, CopyConstructorNull) {
+  VectorValue<double> value{nullptr};
+
+  VectorValue<double> other_value(value);
+  EXPECT_EQ(other_value.get_value(), nullptr);
+}
+
+GTEST_TEST(VectorValueTest, AssignmentOperator) {
+  VectorValue<double> value(BasicVector<double>::Make({1, 2, 3}));
+  VectorValue<double> other_value(BasicVector<double>::Make({4, 5, 6}));
+
+  value = other_value;
+  EXPECT_EQ(4, value.get_value()->get_value().x());
+  EXPECT_EQ(5, value.get_value()->get_value().y());
+  EXPECT_EQ(6, value.get_value()->get_value().z());
+}
+
+GTEST_TEST(VectorValueTest, AssignmentOperatorSubclass) {
+  VectorValue<double> value(MyVector2d::Make(1, 2));
+  VectorValue<double> other_value(BasicVector<double>::Make({5, 6}));
+  other_value = value;
+  value.get_value()->set_value(Eigen::Vector2d(3, 4));
+
+  const MyVector2d* const vector =
+      dynamic_cast<const MyVector2d*>(other_value.get_value());
+  ASSERT_NE(vector, nullptr);
+  EXPECT_EQ(vector->get_value()(0), 1);
+  EXPECT_EQ(vector->get_value()(1), 2);
+}
+
+GTEST_TEST(VectorValueTest, AssignmentOperatorNull) {
+  VectorValue<double> value{nullptr};
+  VectorValue<double> other_value(BasicVector<double>::Make({4, 5, 6}));
+
+  value = other_value;
+  other_value = VectorValue<double>{nullptr};
+  EXPECT_EQ(4, value.get_value()->get_value().x());
+  EXPECT_EQ(5, value.get_value()->get_value().y());
+  EXPECT_EQ(6, value.get_value()->get_value().z());
+  EXPECT_EQ(other_value.get_value(), nullptr);
+}
+
+GTEST_TEST(VectorValueTest, AssignmentOperatorSelf) {
+  VectorValue<double> value(BasicVector<double>::Make({4, 5, 6}));
+  value = value;
+  EXPECT_EQ(4, value.get_value()->get_value().x());
+  EXPECT_EQ(5, value.get_value()->get_value().y());
+  EXPECT_EQ(6, value.get_value()->get_value().z());
+}
+
+}  // namespace
+}  // namespace systems
+}  // namespace drake
